[TS Demuxer] Very basic mpeg2 indexer
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@4737 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ plugins/ADM_demuxers/MpegTS/ADM_tsAudioProbe.h
@@ -21,7 +21,7 @@ using std::vector;
 typedef struct
 {
     WAVHeader header;
-    uint8_t   esID;
+    uint32_t  esID;
 }tsAudioTrackInfo;
 
 typedef vector <tsAudioTrackInfo*> listOfTsAudioTracks;
